refactor(feedback): separate generate userId
