+++ packages/feedback/src/react-components/hooks/useThumbsUp.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect, useRef } from 'react'
-import { v4 as uuidv4 } from 'uuid';
 
+import useUser from './useUser';
 import { getLikes, giveLikes } from '../api'
 
 export default function useThumbsUp() {
-  const [userId, setUserId] = useState(null)
   const [thumbsUp, setThumbsUp] = useState(null)
   const originalThumbUpRef = useRef(null)
+  const { userId } = useUser()
 
   const giveThumbUp = async (thumbUp) => {
     console.log(`send thumbUp '${thumbUp}' to BE`);
@@ -14,15 +14,15 @@ export default function useThumbsUp() {
     const originalThumbUp = originalThumbUpRef.current
     console.log(originalThumbUp)
     if (thumbUp) {
-      setThumbsUp(thumbsUp => ({
+      setThumbsUp({
         thumbUp: originalThumbUp.thumbUp + 1,
         thumbDown: originalThumbUp.thumbDown,
-      }))
+      })
     } else if (thumbUp === false) {
-      setThumbsUp(thumbsUp => ({
+      setThumbsUp({
         thumbUp: originalThumbUp.thumbUp,
         thumbDown: originalThumbUp.thumbDown + 1,
-      }))
+      })
     } else {
       setThumbsUp(originalThumbUpRef.current)
     }
@@ -40,16 +40,6 @@ export default function useThumbsUp() {
       console.log('error', error)
     }
   }
-
-  useEffect(() => {
-    if (sessionStorage['corvid-19-query-user-id']) {
-      setUserId(sessionStorage['corvid-19-query-user-id'])
-    } else {
-      const userId = uuidv4()
-      setUserId(userId)
-      sessionStorage['corvid-19-query-user-id'] = userId
-    }
-  })
 
   useEffect(() => {
     const getThumbsUp = async () => {
